chore: remove fetchAllBuckets flagging code
