+++ src/settings/components/LoadDataTabbedPage.tsx
@@ -12,7 +12,6 @@ import {getOrg} from 'src/organizations/selectors'
 
 // Types
 import {AppState} from 'src/types'
-import {isFlagEnabled} from 'src/shared/utils/featureFlag'
 
 interface ComponentProps {
   activeTab: string
@@ -46,7 +45,7 @@ const LoadDataTabbedPage: FC<Props> = ({activeTab, orgID, children}) => {
 
 // this function returns whether the page should be allowed to scroll or not based on the featureFlag enabled and the current tab the user is on.
 const shouldPageBeScrollable = (activeTab: string): boolean => {
-  if (activeTab === 'buckets' && isFlagEnabled('fetchAllBuckets')) {
+  if (activeTab === 'buckets') {
     return false
   }
   if (activeTab === 'tokens') {
